[Go] change RunFlow and StreamFlow to methods
Make Run and Stream methods on Flow instead of top-level functions.
