+++ go/genkit/genkit.go
@@ -17,7 +17,6 @@ package genkit
 
 import (
 	"context"
-	"errors"
 	"net/http"
 
 	"github.com/firebase/genkit/go/core"
@@ -93,55 +92,6 @@ func DefineStreamingFlow[In, Out, Stream any](
 func Run[Out any](ctx context.Context, name string, f func() (Out, error)) (Out, error) {
 	return core.InternalRun(ctx, name, f)
 }
-
-// RunFlow runs flow in the context of another flow. The flow must run to completion when started
-// (that is, it must not have interrupts).
-func RunFlow[In, Out, Stream any](ctx context.Context, flow *core.Flow[In, Out, Stream], input In) (Out, error) {
-	return core.InternalRunFlow(ctx, flow, input)
-}
-
-// StreamFlowValue is either a streamed value or a final output of a flow.
-type StreamFlowValue[Out, Stream any] struct {
-	Done   bool
-	Output Out    // valid if Done is true
-	Stream Stream // valid if Done is false
-}
-
-// StreamFlow runs flow on input and delivers both the streamed values and the final output.
-// It returns a function whose argument function (the "yield function") will be repeatedly
-// called with the results.
-//
-// If the yield function is passed a non-nil error, the flow has failed with that
-// error; the yield function will not be called again. An error is also passed if
-// the flow fails to complete (that is, it has an interrupt).
-// Genkit Go does not yet support interrupts.
-//
-// If the yield function's [StreamFlowValue] argument has Done == true, the value's
-// Output field contains the final output; the yield function will not be called
-// again.
-//
-// Otherwise the Stream field of the passed [StreamFlowValue] holds a streamed result.
-func StreamFlow[In, Out, Stream any](ctx context.Context, flow *core.Flow[In, Out, Stream], input In) func(func(*StreamFlowValue[Out, Stream], error) bool) {
-	return func(yield func(*StreamFlowValue[Out, Stream], error) bool) {
-		cb := func(ctx context.Context, s Stream) error {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if !yield(&StreamFlowValue[Out, Stream]{Stream: s}, nil) {
-				return errStop
-			}
-			return nil
-		}
-		output, err := core.InternalStreamFlow(ctx, flow, input, cb)
-		if err != nil {
-			yield(nil, err)
-		} else {
-			yield(&StreamFlowValue[Out, Stream]{Done: true, Output: output}, nil)
-		}
-	}
-}
-
-var errStop = errors.New("stop")
 
 // StartFlowServer starts a server serving the routes described in [NewFlowServeMux].
 // It listens on addr, or if empty, the value of the PORT environment variable,
