+++ go/samples/menu/s05.go
@@ -125,7 +125,7 @@ func setup05(ctx context.Context, gen, genVision *ai.Model) error {
 
 	genkit.DefineFlow("s05_visionMenuQuestion",
 		func(ctx context.Context, input *menuQuestionInput) (*answerOutput, error) {
-			menuText, err := genkit.RunFlow(ctx, readMenuFlow, struct{}{})
+			menuText, err := readMenuFlow.Run(ctx, struct{}{})
 			if err != nil {
 				return nil, err
 			}
@@ -134,7 +134,7 @@ func setup05(ctx context.Context, gen, genVision *ai.Model) error {
 				MenuText: menuText,
 				Question: input.Question,
 			}
-			return genkit.RunFlow(ctx, textMenuQuestionFlow, questionInput)
+			return textMenuQuestionFlow.Run(ctx, questionInput)
 		},
 	)
